moveing backbutton logic to top level and reusing for other use cases.  provide additional error handling as well
