+++ apps/web/src/app/not-found.tsx
@@ -1,11 +1,11 @@
+import { BackButton } from "@/app/BackButton";
+
 export default function NotFound() {
   return (
     <div className="flex flex-col items-center justify-center min-h-[60vh] text-center">
       <h1 className="text-4xl font-bold text-cream mb-2">404</h1>
       <p className="text-cream-dim mb-6">Page not found.</p>
-      <a href="/" className="text-accent hover:underline">
-        Back to search
-      </a>
+      <BackButton className="text-accent hover:underline" />
     </div>
   );
 }
